Create a class named ‘Animal’ which includes methods like eat() and sleep().
Create a child class of Animal named ‘Bird’ and override the parent
class methods. Add a new method named fly().

Create an instance of Animal class and invoke the eat and sleep methods
using this object.

Create an instance of Bird class and invoke the eat, sleep and fly
methods using this object.